feat: 조회 로직 구현 및 Join fetch 적용으로 최적화 / API 구현

+++ backend/src/main/java/com/deliveranything/domain/delivery/dto/response/DeliveredSummaryResponseDto.java
@@ -5,9 +5,9 @@ import lombok.Builder;
 
 @Builder
 public record DeliveredSummaryResponseDto(
-    Long thisWeekDeliveredCount,
-    Integer waitingSettlementAmount,
-    Integer completedSettlementAmount,
+    Integer thisWeekDeliveredCount,
+    Long waitingSettlementAmount,
+    Long completedSettlementAmount,
     CursorPageResponse<DeliveredDetailsDto> deliveredDetails
 ) {
